Corrections and Tests
Updated bubbleMerge with refactored binarySearch

Noticed some matches from longestCommonSubstring will intersect with others in terms of AREA, but seems like a dud, will not pursue this.

AddedBubbleMergeTestCase for LongestCommonSubString, a dud, will not pursue this.

played with longestcommonstring testcase, to see effect of swapping inputs

latexparsertest ids got swapped, but AST was isomorphic to previous results, so updated results.