TESTED: IMPLEMENTATION and CORRECTION and PLANNING
removed incorrect apsp, planned 'correct' algo for apsp, but did not implement, since i do not need it now...

implemented a easier to understand smallcyclefinder and seems to work with smalltest, commented out old plan

corrected minimumspanningtree, used Kruskal(sorry about it!), realised that its actually bFSTree?. Did not use it in the end.

implemented BFSDFSTree in spanningtree, since they are the same thing, and somehow, the minimumspanningtree might be faster to run, will try a few example circuits.

created a fake file for allpairshortestpathtest, but test is not correct

tested small cyclefinder seems to work, but a bit slow, might be no pycache, and was compiling (not sure)

replaced uuid with integers before sending to backend, easier to debug on backend.

stored the uuid to integers mapping on front end class

checked new networkGraph on views.py backend

some security configuration on settings

will use the smallCycleFinder with id__type to generate KVL equations, by assuming all non-wire components on cycles to have variable voltage and add up to zero, for this test we should have 2 equations, replanned the equation_patterns order, and thought about specifics