TESTED: REFACTOR and PLAN
Made getVariable a classmethod, so that it could be easily used in view.py for frontend to transmit the independent and dependent_variable.

Planned to check the <<3000>> example, but it was more difficult because the independent is not a component, but a sum of resistance.

Moved circuit code, like wire, to circuit.js from piece.js

Refactored animation wireframe from dc_series example to circuit.js, so less repeats as we put in more examples from <<3000>>

Allowed frontend to request for specific independent and list of dependent_variables, instead of hardcode