TESTED: FEATURE and PLAN and TEST
animation seems to be running as intended, except equation_variables highlights variable wrongly

tested animation on only the 2 simple circuits

added equationFinderDisplayName to all the equationFinder, for future display

added associatedComponentIdList to all the equationFinder, for highlighting components associated with the equation found

in the ~Component.js. there are also of other meshes like the touchingFaces that need not be highlighted, so added a mainMeshUUID for meshes that needs to be highlighted

planned to debug the wrong highlight of some variables in some equations, i suspect, its because i associated the latexMesh wrongly with the nodeId

included another version of MathJax, CHTML, and reassigned the window names of the different MathJax

included MathJaxCHTML to use the stylised_unicode mapping in it, to get the nonstylised character of the stylised_unicode tagged to each latexMeshLetter

removed axisHelper on main.js

added an Actor class to contain common functions for Components and LatexMesh, might also need it for later mechanical|molecular simulations, all ~Component.js now extend from Actor.js

copied the Animation.js structure from ChatGPT, and it worked well

used MathJaxCHTML mapping to associate latexMesh_meshUUID to the actual letters in the latexStr, and then grouped the meshUUID into the actual variables of the latexStr, but i think i made some mistake here. This is done in the LatexMeshCreator, and then used to get the latexMeshUUID for highlighting

because of above changes to ~Component.js, i had to regenerate the all the children components

also had to do the same changes manually to the Wire.js

wrote createComponent, now all the components must be created through circuit.js, so that we can register all the components as actors

also corrected wire function in circuit.js because of createComponent

also corrected all the initialisation of components in the 2 simple circuits

implemented findEquation animation feature in the 2 simple circuits for testing

planned to rely more on the <<foundations>> book instead of <<3000>>, so i made a dummy circuit file, but have not filled it up yet

added many helper functions to piece.js, all for the purpose of creating animation

refactored the view.py findEquation, so that it returns the information need by frontend to create the animation.