TESTED: PLANNING and FEATURE and TEST
planned to ameriolate and fix remaining bugs before YTU

audio and video roughly aligned, but only tested for two_resistor_parallel,and not yet test if recorder still works

made a simple wave player, to suit the new process where we generate a new wavFile for every sentence, instead of relying on pause and start of prerecorded timings

orbitTurn seems to be called many times, tried to prevent that, but to no avail, for now, i do not wish to hack it any further.

removed the pyttsx3 package, because we integrated it into this project

a different wavGenFunc, because we want to keep the old one, just in case.

unit test for new wavGenFunc

removed some unnecessary comments