UNTESTED: TRIED_TO_TEST
tried to create test and then realised that some variables cannot be remove to get what we want because sometimes, we have to use Ohms_law and sometimes we have to use timing_model in OhmLaw place. And i have not figured out what is sometimes. This sometimes cannot be component dependant, because filters have resistors (which mostly use OhmLaw) together with capacitors (which mostly use timingModels)